feat: setup test with jest

+++ src/migration.ts
@@ -1,4 +1,5 @@
 import 'reflect-metadata';
+import readline from 'readline';
 import { Database } from '@config/db';
 import { logger } from '@common/logger';
 import { argv } from 'process';
@@ -14,10 +15,34 @@ export const handler = async (event: string): Promise<string> => {
       await dataSource.runMigrations();
       break;
     case 'down':
-      await dataSource.undoLastMigration();
+      var rl = readline.createInterface(process.stdin, process.stdout);
+      console.log(
+        '\x1b[31m%s\x1b[0m',
+        '!!!!!!!!!!!!!!!!!!! DANGEROUS ACTION !!!!!!!!!!!!!!!!!!!',
+      );
+      rl.question('Still done? [yes]/no: ', async (answer) => {
+        if (answer !== 'yes') {
+          console.log('Aborted');
+          process.exit(0);
+        } else {
+          await dataSource.undoLastMigration();
+        }
+      });
       break;
     case 'drop':
-      await dataSource.dropDatabase();
+      var rl = readline.createInterface(process.stdin, process.stdout);
+      console.log(
+        '\x1b[31m%s\x1b[0m',
+        '!!!!!!!!!!!!!!!!!!! DANGEROUS ACTION !!!!!!!!!!!!!!!!!!!',
+      );
+      rl.question('Still done? [yes]/no: ', async (answer) => {
+        if (answer !== 'yes') {
+          console.log('Aborted');
+          process.exit(0);
+        } else {
+          await dataSource.dropDatabase();
+        }
+      });
       break;
     case 'show':
       await dataSource.showMigrations();
